backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.7.1
Flask-Cors==5.0.0
python-dotenv==1.0.1
Werkzeug==3.1.3
Pillow==11.1.0
email-validator==2.2.0
psycopg2-binary
gunicorn
